pyproject.toml
[tool.poetry]
name = "book-api"
version = "0.1.0"
description = ""
authors = ["Shane Bell <being.for.itself@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.6" }
sqlmodel = "^0.0.22"
exceptiongroup = "^1.2.2"
async-timeout = "^5.0.1"
sqlalchemy-libsql = "^0.1.0"
pg8000 = "^1.31.2"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
scalar-fastapi = "^1.0.3"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pipenv-poetry-migrate = "^0.5.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# dev = "fastapi dev ./book_api/main.py"
